requirements.txt
asgiref==3.4.1
click==8.0.1
dataclasses==0.8
ecdsa==0.17.0
fastapi==0.68.1
h11==0.12.0
importlib-metadata==4.8.1
pkg-resources==0.0.0
pyasn1==0.4.8
pydantic==1.8.2
PyJWT==2.1.0
python-dotenv==0.19.0
python-jose==3.3.0
rsa==4.7.2
selenium==3.141.0
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
zipp==3.6.0
